PosService, Payment ve kontrolleri eklendi.

+++ rentAcar/src/main/java/com/turkcell/rentAcar/business/dtos/invoice/GetInvoiceDto.java
@@ -21,7 +21,7 @@ public class GetInvoiceDto {
 
 	private LocalDate returnDate;
 
-	private int totalDay;
+	private long totalDay;
 
 	private double rentTotalPrice;
