+++ rentAcar/src/main/java/com/turkcell/rentAcar/business/dtos/invoice/ListInvoiceDto.java
@@ -20,7 +20,7 @@ public class ListInvoiceDto {
 
 	private LocalDate returnDate;
 
-	private int totalDay;
+	private long totalDay;
 
 	private double rentTotalPrice;
